[Fix] ASIO: Allow buffer sizes smaller than 32 samples. [Fix] ASIO: Actually honor user-selected latency value for devices that require power-of-2 buffer sizes instead of blindly falling back to peferred size (broken since at least r37 ). [Ref] ASIO: Rewrite buffer size selection code to be more tolerant about broken ASIO drivers.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@3267 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ sounddev/SoundDeviceASIO.h
@@ -34,7 +34,7 @@ class CASIODevice: public ISoundDevice
 	enum { ASIO_MAX_CHANNELS=4 };
 protected:
 	IASIO *m_pAsioDrv;
-	UINT m_nAsioBufferLen;
+	long m_nAsioBufferLen;
 	UINT m_nAsioSampleSize;
 	bool m_Float;
 	bool m_CanOutputReady;
